--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/annotation/InputMissingDeprecated2.java
@@ -1,14 +1,0 @@
-package com.puppycrawl.tools.checkstyle.checks.annotation;
-
-public class InputMissingDeprecated2 {
-
-
-    /**
-     * @deprecated
-     * 
-     * @param comment
-     */
-    public void method(){
-    }
-
-}
