+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/annotation/missingdeprecated/InputMissingDeprecatedMethod.java
@@ -1,0 +1,14 @@
+package com.puppycrawl.tools.checkstyle.checks.annotation.missingdeprecated;
+
+public class InputMissingDeprecatedMethod {
+
+
+    /**
+     * @deprecated
+     * 
+     * @param comment
+     */
+    public void method(){
+    }
+
+}
